Consolidate data for AudioContext constructor parameters (#11364)
The latencyHint and sampleRate parameter support was represented in two
places. Consolidate into the subentries of the constructor and name them
per current guideline.

Any existing version numbers were assumed to be correct and copied over,
and the some additional data was mirrored.

The Safari data is new and based on this change:
https://trac.webkit.org/changeset/264342/webkit
https://trac.webkit.org/browser/webkit/trunk/Source/WebCore/Configurations/Version.xcconfig?rev=264342

Although latencyHint is in WebKit's IDL, there's FIXME and a bug to
actually use it, which is linked.