Rename/move entries for Selection method parameters
Rename two of them to be more consistent with other parameter entries in
BCD and importantly to have underscores in the name.

The subfeatures for modify() are moved to collapse(). The modify()
method doesn't have a node or offset parameter, and this data originates
from https://github.com/mdn/browser-compat-data/pull/1738 where
there appears to have been a mishap while splitting a "params"
subfeature for collapse().