Drop Sanitizer.p.sanitizeToString
https://github.com/WICG/sanitizer-api/commit/f1b12c1
(https://github.com/WICG/sanitizer-api/pull/99) completely removed
Sanitizer.prototype.sanitizeToString from the HTML Sanitizer API spec.

It never shipped in any browsers other than Firefox, and in Firefox it
only ever shipped behind a flag.

So there’s no value to developers in keeping it around forever in BCD
and in MDN. It was never a cross-browser feature of the platform.