Remove obsolete worker.close() and related entries (#11557)
There are close() methods on both
DedicatedWorkerGlobalScope and SharedWorkerGlobalScope in BCD and MDN
already:
https://developer.mozilla.org/en-US/docs/Web/API/DedicatedWorkerGlobalScope/close
https://developer.mozilla.org/en-US/docs/Web/API/SharedWorkerGlobalScope/close

There's a note here that explains how this came about:
https://developer.mozilla.org/en-US/docs/Web/API/SharedWorkerGlobalScope/close#browser_compatibility

However, we need not keep this data or even have notes about it:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines.md#apis-moved-on-the-prototype-chain

Furthermore, there is no close event or onclose property. It's not clear
when this was removed from browsers, but the onclose property is at
least not supported anywhere today.

This also updates some existing data:

Verified supported in Opera 10.6:
http://mdn-bcd-collector.appspot.com/tests/api/DedicatedWorkerGlobalScope/close
http://mdn-bcd-collector.appspot.com/tests/api/SharedWorkerGlobalScope/close

Firefox for Android data was also bumped to 29, it's implausible that it
was supported much earlier on Android. It's probably the result of
copying data from DedicatedWorkerGlobalScope, which is marked as
supported since firefox for Android 4.