Fix data for focus() and blur() methods (#11637)
For SVGElement.blur, the data for focus() was copied as a starting
point. Then http://mdn-bcd-collector.appspot.com/tests/api/SVGElement/blur
was tested on these browser versions:
 - Chrome 15 on Windows 7
 - Edge 16+17 on Windows 10
 - Firefox 50+51 on Windows 10
 - Opera 12.16 on Windows 7 (not supported)
 - Safari 4 on Mac OS X 10.6.8

Not all of the data can be inferred from the results, but it's
consistent with the results.

The HTMLSelectElement entries are removed, since they're not in any spec
and there are no MDN pages for them. Support on HTMLSelectElement may
have slightly preceded focus() and blur() on HTMLElement, but it's so
long ago that it's not worth capturing with notes.