Add all browsers versions for FileReader API
This PR adds real values for all browsers for the `FileReader` API.  The data for the events was copied from their event handler counterparts.  Data for the other features was mirrored from Chrome to Chrome Android and Samsung Internet.  Finally, any mentions of Safari 6.1 was corrected to Safari 7, since we are removing the 6.1 backport data.