Remove KeyboardEvent.which and MouseEvent.which
UIEvent.which is already in the data, so update that to reflect the
partial support from when `which` was on KeyboardEvent and MouseEvent.

The existing data from MouseEvent is not perfectly preserved, in
particular the early Opera and Safari versions have not been verified,
so instead ranges are used in the merged data.