Remove irrelevant documentURI.readonly feature
The original behavior, being able to set document.documentURI, is a
superset of the current behavior, and there isn't anything worth noting
to users of the API today. The mutable behavior, if seen as its own
feature, has been removed for long enough that it's irrelevant.

https://mdn-bcd-collector.appspot.com/tests/api/Document/documentURI/readonly
was used to determine that documentURI was made readonly in Edge 17,
the only part of the data that could make this change <2 years.