Add Safari versions for api.Element.animate/getAnimations (#12275)
This PR adds real values for Safari (Desktop and iOS/iPadOS) for the `animate` and `getAnimations` members of the `Element` API.  It turns out the "CSS Animations via Web Animations" flag had no effect upon these properties.  The data was obtained by going through SauceLabs, checking each individual version, turning on the flag (if it was available), and seeing what the first version that supported the property was.

Also fix getAnimations() data which is no longer behind a flag.