Add Firefox versions for api.Text.replaceWholeText (#12520)
This PR adds real values for Firefox and Firefox Android for the `replaceWholeText` member of the `Text` API.  This is just setting a version number for this feature, because this feature is queued for removal as it is.  This is just to make sure that it doesn't get in the way of the real value percentage in the meantime.