Drop the spec URL for FontFaceSet»entries()
There is no #dom-fontfaceset-set-entries anchor in the
https://drafts.csswg.org/css-font-loading/ spec — because the spec
itself doesn’t define an entries() method; instead, because the
interface it annotated with “setlike”, the WebIDL spec, at
https://heycam.github.io/webidl/#idl-setlike, defines it as having an
entries() method.

But we have previously intentionally chosen to not link to the WebIDL
spec from BCD for these “setlike”-annotated interfaces.

So we just need to drop the spec URL for this.