Remove irrelevant notes about TextEncoder (#12979)
There was once a parameter to use other encodings than UTF-8, but this
was now long ago, and this isn't something developers need to take into
account when supporting old browsers, since new TextEncoder() without
argument works just fine. If the parameter were represented as a
subfeature, it would be removed as irrelevant under this guideline:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines.md#removal-of-irrelevant-features

Also remove the notes about Firefox 18. No details are provided, and the
fact that some parts of the API shipped later is already captured. Basic
usage like this was confirmed to work in Firefox 18:
```js
var e = new TextEncoder();
console.log(e.encode('hallå'));
```

That the feature shipped in Chrome 38 + Firefox 18 was also reconfirmed
using this test:
http://mdn-bcd-collector.appspot.com/tests/api/TextEncoder