Remove api.isSecureContext.considers_window_opener
This PR removes `api.isSecureContext.considers_window_opener` from BCD.  This feature is marked as standard and experimental, but looking through the spec I cannot find any mention of an `opener` or anything relevant for this property.  It's possible this was added to a draft version of the spec, added in Firefox, but then removed from the spec.

Additionally, there is no documentation on on the MDN page that discusses what this feature means exactly.