Set most of <feGaussianBlur> to supported in all browsers
Support goes very long back and can also be confirmed by the data for
the SVGFEGaussianBlurElement interface:
https://github.com/mdn/browser-compat-data/blob/main/api/SVGFEGaussianBlurElement.json

The exact versions aren't copied, however, because this far back it's
possible that the elements were supported before the DOM API for them.

This is motivated by fixing an apparent Chrome-only omission of `in` and
`stdDeviation`, if null is interpreted as false which I did for
convenience in a spreadsheet exercise to find such Chrome-only
omissions.

The `edgeMode` data isn't updated, because it's seemingly Safari-only,
which seems wise to confirm before copying the data, and that would be a
lot of work.