Remove the "type and listener parameters are optional" feature (#13143)
This is the inversion of the non-standard "feature" that the following
don't throw an exception but instead do nothing:
```js
window.addEventListener()
window.addEventListener('load')
```

That "feature" has been removed from all browsers, confirmed by manually
testing whether `window.addEventListener('load')` throws.

Notably, it did throw in Edge 18. Exactly when the parameters were made
required in Chrome, Firefox and Safari hasn't been tested. It was
probably long ago, but ultimately it doesn't matter, since the "feature"
here is not a feature at all, just a quirk that was only important at
the moment it was being fixed, when it could have broken sites.