Add IE/Edge versions for api.Document.createElementNS.options_parameter (#13192)
This PR adds real values for Internet Explorer and Edge for the `createElementNS.options_parameter` member of the `Document` API.  The data for `createElement.options_parameter` was set to `false` for IE and pre-Chromium Edge in https://github.com/mdn/browser-compat-data/pull/4975.  I doubt that this would be supported in `createElementNS` if `createElement` doesn't support it.