Add WebView versions for MouseEvent API
This PR adds real values for WebView Android for the `MouseEvent` API.  The data was mirrored from Chrome/Safari and determined based upon the WebKit version.  I'm not able to set the version added to "≤37" because the version removed is "37" -- the linter will fail due to the changes made in #10710.