Remove calc()'s "<color> value support" entry
This entry represents support for calc() inside color functions like
rgb(), and can be tested with `rgb(0, calc(64 + 64), 0)`:
http://software.hixie.ch/utilities/js/live-dom-viewer/?saved=9798

The data for Firefox is correct, this didn't work before Firefox 59.

However, the data for Chrome and Edge is wrong, this was supported since
Chrome 16 (with `-webkit-calc()`) and at least since Edge 15.

Since this concerns only Firefox, make it a note instead.