Remove api.Range.insertNode.collapsed_ranges (#13424)
It's not clear what this sub-feature means from just the description,
and all we have is a guess:
https://github.com/mdn/browser-compat-data/pull/13075#issuecomment-954756548

The data matches the parent feature except for Firefox and Safari. This
probably means there's *something* that doesn't work the same before
Firefox 14 and Safari 4. This is far enough in the past that it's not
worth trying to salvage this sub-feature with more research.