Note that Firefox-for-Android supports ::part since v72 (#13628)
* Note that Firefox-for-Android supports ::part since v72

::part shipped in Firefox (all platforms) in version 72, in https://bugzilla.mozilla.org/show_bug.cgi?id=1559074

It looks like Firefox-for-Android was missed when this json file was updated to reflect that.  This commit addresses that and marks it as supported.

(Note: the Firefox-for-Desktop "behind-a-pref" snippet for versions 69-72 is likely accurate for Firefox-for-Android, too, but I'm not bothering to copy that over, for brevity & because it's unlikely to matter at this point.)

* Update Firefox for Android version number to 79, for ::part's first-supported-version

The API was added in Gecko 72, but that didn't ship in a Firefox release on Android until version 79.