Remove the requestAnimationFrame() "Return value" subfeature (#13694)
This is already captured by the cancelAnimationFrame() entry.

The return value of requestAnimationFrame() is an integer which can be
used to cancel the callback with cancelAnimationFrame(). Judging by the
data, this entry must have been added to capture that
mozRequestAnimationFrame() returned undefined in Firefox 4-10, but this
by itself is not useful information.

Also, most readers could probably not decipher what this entry means.