Add Safari versions for RTCRtpReceiver API (#13813)
This PR adds real values for Safari (Desktop and iOS/iPadOS) for the `RTCRtpReceiver` API.  Honestly speaking, I have no clue how to test support for these features, and they're the last ones remaining that have true/nnull values.  This PR is simply filling in version numbers for the sake of completing the version numbers, and rely on the community to make changes if needed.