Update support for flex-basis property values
Support for content and min/max/fit-content shipped in Chrome 94:
https://storage.googleapis.com/chromium-find-releases-static/b3f.html#b3fc034c85b71d3575333731879d82e2e729d0e9

-moz-fit-content was unprefixed in Firefox 94:
https://bugzilla.mozilla.org/show_bug.cgi?id=1732759

This was also confirmed using
`CSS.supports('flex-basis', 'fit-content')` in Firefox 93 and 94 and
support for the prefixed form in Firefox 22 was confirmed using
`CSS.supports('flex-basis', '-moz-fit-content')`.

Lack of support for fit-content in Edge and Safari was confirmed using
`CSS.supports('flex-basis', 'fit-content')`.