Update the javascript/operators/await spec URL
The spec URLs for the ES spec should have not just an anchor but also a
filename for the multipage HTML filename the anchor is in. (Browsers do a
client-side redirect of the links, so that the user ends up in the right
file anyway. But having the filename avoids that unnecessary redirect.