Add Chromium versions for javascript.builtins.Atomics.notify
This PR adds real values for Chromium (Chrome, Opera, Samsung Internet, WebView Android) for the `notify` member of the `Atomics` JavaScript builtin.  The data for Samsung Internet was mirrored from Chrome, but it's set to `false` because it was added and removed from Chrome before a Samsung Internet release came out.