Replace hyphens with underscores in some subfeature names
Although not enforced by any lint, the dominant convention in BCD is to
use underscores in subfeatures describing some behavior, as opposed to
the directly observable API surface. This makes it more obvious that
it's not the name of an API, and can be mostly ignored when working on
mdn-bcd-collector.

Rename some of the subfeatures that currently use hyphens for no very
good reason. Some still remain that are named by some code name, like
"default_same-origin", but the remaining names include at least one
underscore, which is good enough for filtering.