Update Chromium versions for MediaQueryList
Original source: https://github.com/mdn/browser-compat-data/pull/3255

https://chromium.googlesource.com/chromium/src/+/5fc555b830a9b5b1536e242d8754d7e507bcc75a
was used as the source, which is the right commit but it wasn't first in
Chrome 45 but earlier.

At the time, the MediaQueryList interface had [NoInterfaceObject], but
because MediaQueryListEvent was implemented/exposed at the same time, we
can assume that the versions should match. Chrome 39 was confirmed with
this test:
https://mdn-bcd-collector.appspot.com/tests/api/MediaQueryListEvent

The data for MediaQueryListEvent already matches.

Part of https://github.com/mdn/browser-compat-data/issues/7844.