Remove three_value_syntax features
This has been removed in all engines and browser releases for more than
two years. The WebKit removal isn't reflected in the data, but was here:
https://bugs.webkit.org/show_bug.cgi?id=189142

css.properties.transform-origin.three_value_syntax remains and is
different, so fix a typo in an example referring to it.

Fixes https://github.com/mdn/browser-compat-data/issues/8964.