Mark Push API as shipped in Firefox for Android 48 (#16323)
The "Push enabled by default." notes don't really make sense as used,
since it means the API wasn't supported before that. So change all parts
of the API to be enabled in Firefox for Android 48.

The version wasn't confirmed with testing, but per
https://bugzilla.mozilla.org/show_bug.cgi?id=1278435 the API was enabled
separately for Android, and 48 is the earliest affected version, along
with 49 and 50.