Allow optional `implementation_url` in support data (#16415)
* schema: Allow optional implementation_url

This change updates the compat-data schema and docs to allow per-browser
optional implementation_url values to be specified in the support data
for any browser; detailed description:

> An optional changeset URL or commit URL for the change which implemented
> the feature in the source code for the associated browser; e.g. a
> https://trac.webkit.org/changeset/ URL for a changeset whose subject
> line is in the form *"Enable [feature name/description]."* or *"Enable
> [feature name/description] by default."*. The presence of an
> `implementation_url` value indicates that the associated browser has
> spec-conformant support for the feature in its source code. Therefore,
> an `implementation_url` value can be added even if the feature hasn't
> yet shipped in a stable release, but instead only in a Nightly, Canary,
> or Technology Preview release — or even if the feature hasn't yet
> shipped in any release at all, but has only landed in the source code
> for the associated browser.

https://github.com/mdn/browser-compat-data/issues/6896#issuecomment-707928572
and https://github.com/mdn/browser-compat-data/issues/6896#issuecomment-708876962
are where the idea for this change arose. In summary, the rationale is to
have a record in BCD of what changeset landed support for a feature in a
particular browser’s codebase. Among other use cases, we have downstream
consumers who care only whether or not there’s an implementation of the
feature in a particular browser engine — even if that implementation
hasn’t yet actually shipped in a major release.

* Make implementation_url clearly be changeset URLs

This change updates the compat-data schema docs to provide guidance that
implementation_url values will be https://trac.webkit.org/changeset/,
https://hg.mozilla.org/mozilla-central/rev/, or https://crrev.com/ URLs.
The change also updates the schema itself to constrain the allowed value
to a pattern that matches URLs in one of those forms.

* Update schemas/compat-data-schema.md

Co-authored-by: Daniel D. Beck <daniel@ddbeck.com>

* Format

* Allow for bug URLs as implementation URL

* Shorten chromium.googlesource.com -> crrev.com

* Update descriptions

* Allow implementation URL value to be an array

* Update documentation

* Update mirroring script to account for new property

* Update description

* implementation_url -> impl_url

Co-authored-by: Michael[tm] Smith <mike@w3.org>
Co-authored-by: Daniel D. Beck <daniel@ddbeck.com>