SharedWorkerGlobalScope needs version_removed for Safari
I suspect the current data here comes from the BCD Collector, which
will have the data values for later Safari versions as null (as it
cannot check whether SharedWorkerGlobalScope is supported, as it
cannot access it, as it doesn't have SharedWorkers), hence the
existing data omits the fact that Safari removed support. The data
here can be compared with that in ./SharedWorker.json, which does
indeed document the removal.