Add more spec URLs for the “alternative stylesheet” feature
https://html.spec.whatwg.org/multipage/#rel-alternate alone as a spec
URL does not adequately capture the scope of the “alternative
stylesheets” feature.

https://developer.mozilla.org/docs/Web/CSS/Alternative_style_sheets#specifications
has five spec citations in total — which we’re currently using the
spec-urls YAML frontmatter metadata key to record.

But it would be better to have those spec URLs in BCD itself — then we
can switch the MDN article to just using the browser-compat key, so that
the spec URLs get pulled in from BCD too.