Add data for attributes of the HTML input element (#16493)
Fixes https://github.com/mdn/browser-compat-data/issues/3157

This change adds data for content attributes of the HTML input element
(in contrast to HTMLInputElement IDL attributes, which we already have
data for).

Otherwise, without this change, among the problems we have is that we
a number of MDN articles have broken BCD tables; for example:

- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/accept#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/capture#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/disabled#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/max#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/maxlength#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/min#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/minlength#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/pattern#browser_compatibility
- https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/readonly#browser_compatibility

Also, this change allows us to store the spec URL for these features
in BCD; otherwise we need to use the spec-url frontmatter metadata key
in MDN to store the spec URLs. Storing them in BCD is preferable,
because that gives all consumers of BCD data access to the URLs.