Add resize event entry for HTMLVideoElement
This event is fired when videoWidth or videoHeight changes, so put it on
HTMLVideoElement where those properties are. In the Blink
implementation, firing the event is explicitly conditional on the
element being an HTMLVideoElement, not an HTMLAudioElement.

The same versions as for videoWidth and videoHeight are used without
testing to confirm.