Bump all IndexedDB Safari versions to 8 (was 7)
For iOS, this can be confirmed by testing 7 and 8 on BrowserStack:
https://mdn-bcd-collector.appspot.com/tests/api/indexedDB

For desktop, support can be confirmed using the same test in Safari
7.1.8 on OS X 10.9.5. BrowserStack doesn't have Safari 7 for OS X, so
that can't be tested.

This tweet also supports the notion that it was in 7.1:
https://twitter.com/daleharvey/status/456759068053356544

This is also what https://caniuse.com/indexeddb says.

Since BCD doesn't have Safari 7.1, instead use 8, which is the closest
release in terms of date. This avoid the probably-false claim that it
was supported in Safari 7.