Make notes around event handler properties more consistent (#17061)
This uses the wording from the guideline:
https://github.com/mdn/browser-compat-data/blob/main/docs/data-guidelines.md#dom-events-eventname_event

Extra bits about `addEventListener()` are left alone.

Links to https://crbug.com/868224 are removed except for the remaining
case of onanimationcancel. It was incorrectly linked for non-Chromium
browsers in many places as well.

For SVGAnimationElement, the notes about the event handler property name
are removed, this is already shown by the MDN pages.