Update Safari support for CryptoKey (#16982)
This matches the data for crypto.subtle.generateKey() and importKey(),
the APIs that create CryptoKey instances.

Note that this is probably wrong, but done for consistency.

The initial implementations of importKey() and generateKey() were in
WebKit trunk version 538.5:
https://github.com/WebKit/WebKit/commit/a834b36c558ddbefad9fad00a77d2732a47932ea
https://github.com/WebKit/WebKit/blob/a834b36c558ddbefad9fad00a77d2732a47932ea/Source/WebCore/Configurations/Version.xcconfig
https://github.com/WebKit/WebKit/commit/f656051c47f9c9cbab013808f690b75241897709
https://github.com/WebKit/WebKit/blob/f656051c47f9c9cbab013808f690b75241897709/Source/WebCore/Configurations/Version.xcconfig

That would be Safari/iOS 8.

There's a collector test for this as well:
https://mdn-bcd-collector.appspot.com/tests/api/CryptoKey?exposure=Window

Support is in Safari 7.1 and iOS 8, but not iOS 7. That also suggests
Safari/iOS 8/8 would be more correct. This is a possible follow-up.

Co-authored-by: Queen Vinyl Da.i'gyu-Kazotetsu <vinyldarkscratch@gmail.com>