Update Chromium versions for OVR_multiview2 API
This PR updates and corrects the real values for Chromium (Chrome, Opera, Samsung Internet, WebView Android) for the `OVR_multiview2` API, based upon results from the [mdn-bcd-collector](https://mdn-bcd-collector.appspot.com) project (v6.1.0).

Tests Used: https://mdn-bcd-collector.appspot.com/tests/api/OVR_multiview2

_Check out the [collector's guide on how to review this PR](https://github.com/foolip/mdn-bcd-collector#reviewing-bcd-changes)._

Note: all of the original data is removed because this extension was not supported at all before Chrome 93, as tested in various versions on BrowserStack.  The alternative names only applied to the name of the interface, which are irrelevant in this case.