Update WebKit support for innerHTML/outerHTML/innerText/outerText
innerHTML and innerText was supported on HTMLElement from the beginning of WebKit:
https://github.com/WebKit/WebKit/blob/a063c2b75ddebd1edfe9c54c428d762bdb61f794/WebCore/khtml/ecma/kjs_html.cpp#L908-L911

outerHTML was implemented on 2004-05-05, after the Safari 1.2 release
but long before the Safari 1.3 release, so assume 1.3:
https://github.com/WebKit/WebKit/commit/09b1ec859efb384981e6f6f35e27363d83f78c38

outerText was implemented shortly after, also for 1.3:
https://github.com/WebKit/WebKit/commit/206f85c759065dc3df66c8818ef2eb3682ee7a22

Support for all 4 properties was manually confirmed in Safari 4 by
evaluating `document.body.innerHTML` and similar. This means any
remaining errors are before that time.

All of this was well before the initial releases of WebView Android and
Chrome, so set those versions to 1.

The notes about Node are wrong. innerHTML and outerHTML were originally
on HTMLElement, but that really doesn't matter, since even now when they
are on Element, they throw InvalidStateError for XML documents per spec:
https://w3c.github.io/DOM-Parsing/#ref-for-dom-innerhtml-innerhtml-2
https://w3c.github.io/DOM-Parsing/#ref-for-dom-element-outerhtml-2