Mirror data where it was true and upstream had a real version
This is relatively safe to do for browsers that only been based on one
browser engine, which excludes Edge and Opera. Also, anything with notes
was not mirrored, since that might require a closer review.

These changes were made with an ad-hoc script but would be best reviewed
by comparing the output.