Mark gl.lineWidth() as actually supported (#17245)
The method exists, confirmed with this test in recent Chrome, Firefox
and Safari:
http://mdn-bcd-collector.appspot.com/tests/api/WebGLRenderingContext/lineWidth

https://jsfiddle.net/jkmLae4x/ is a demo and actually the method doesn't
seem to do anything, still drawing a 1px line. There's a warning about
this on MDN already:
https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/lineWidth

Some notes could make sense, but would require deep research to figure
out if the minimum and maximum line width are always 1, or
gl.lineWidth(5) would work in some case.

This effectively reverts this old commit:
https://github.com/mdn/browser-compat-data/commit/3e009d0996879f088fb3cffc77f2cd9f7225ff59

The data before that is trusted, and matches most other features of
WebGLRenderingContext.