Add missing api.OffscreenCanvas.context[lost/restored]_event feature
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser.  This particular PR adds the missing `contextlost_event` and `contextrestored_event` members of the OffscreenCanvas API, populating the results using data from the [mdn-bcd-collector](https://mdn-bcd-collector.appspot.com) project (v6.1.2).

Tests Used:
https://mdn-bcd-collector.appspot.com/tests/api/OffscreenCanvas/contextlost_event
https://mdn-bcd-collector.appspot.com/tests/api/OffscreenCanvas/contextrestored_event

_Check out the [collector's guide on how to review this PR](https://github.com/foolip/mdn-bcd-collector#reviewing-bcd-changes)._