Add missing GPU APIs
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser.  This particular PR adds the missing `FontData` API, populating the results using data from the [mdn-bcd-collector](https://mdn-bcd-collector.appspot.com) project (v6.1.2).

Tests Used:
https://mdn-bcd-collector.appspot.com/tests/api/GPUBuffer
https://mdn-bcd-collector.appspot.com/tests/api/GPUBufferUsage
https://mdn-bcd-collector.appspot.com/tests/api/GPUColorWrite
https://mdn-bcd-collector.appspot.com/tests/api/GPUCommandEncoder
https://mdn-bcd-collector.appspot.com/tests/api/GPUComputePassEncoder
https://mdn-bcd-collector.appspot.com/tests/api/GPUMapMode
https://mdn-bcd-collector.appspot.com/tests/api/GPUQueue
https://mdn-bcd-collector.appspot.com/tests/api/GPURenderPassEncoder
https://mdn-bcd-collector.appspot.com/tests/api/GPUShaderStage
https://mdn-bcd-collector.appspot.com/tests/api/GPUTextureUsage

_Check out the [collector's guide on how to review this PR](https://github.com/foolip/mdn-bcd-collector#reviewing-bcd-changes)._