Add missing WebTransport APIs
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser.  This particular PR adds the missing WebTransport APIs, populating the results using data from the [mdn-bcd-collector](https://mdn-bcd-collector.appspot.com) project (v6.1.2).

Tests Used:
https://mdn-bcd-collector.appspot.com/tests/api/WebTransport
https://mdn-bcd-collector.appspot.com/tests/api/WebTransportBidirectionalStream
https://mdn-bcd-collector.appspot.com/tests/api/WebTransportDatagramDuplexStream
https://mdn-bcd-collector.appspot.com/tests/api/WebTransportError

_Check out the [collector's guide on how to review this PR](https://github.com/foolip/mdn-bcd-collector#reviewing-bcd-changes)._