Update early support for CanvasGradient and CanvasPattern
Support for CanvasPattern is implied by support for ctx.createPattern(),
since that's how a CanvasPattern instance is created. Similarly, support
for CanvasGradient is implied by support for either
createLinearGradient() or createRadialGradient().

The support data for CanvasRenderingContext2D looks reliable based on
the PRs that updated it to the current state:
https://github.com/mdn/browser-compat-data/pull/7465
https://github.com/mdn/browser-compat-data/pull/8666

Update CanvasGradient and CanvasPattern to match. The versions now match
for all browsers except Opera, where some ≤12.1 versions were left alone
rather than trying to confirm the correct versions.