Add subfeatures for menclose@notation
This add subfeatures for each value of the menclose@notation attribute as discussed in [1], except "radical" which is considered irrelevant [2]. This is done by copying existing implementation status of the notation feature.

Additonally, "mirror" is used for Safari iOS (as it has the same MathML support as macOS) and the notes for Safari is removed (as that's redundant information).

[1] https://github.com/mdn/browser-compat-data/pull/17062
[2] https://github.com/mdn/content/pull/21179