Add subfeature for <ms> lquote/rquote attributes
In MathML Core the `<ms>` element behaves like an `<mtext>` [1] and this is how it was implementated in Safari 9 [2] and Chromium.
Firefox still supports a legacy implementation from MathML3 (surround content with quotes, specifiable via attributes) but will be removed in the future [3].
To facilitate comparison of browser support, this commit introduces a subfeature for that legacy implementation.

[1] https://w3c.github.io/mathml-core/#string-literal-ms
[2] https://commits.webkit.org/157302@main
[3] https://bugzilla.mozilla.org/show_bug.cgi?id=1793387