Add `@import layer()`
This adds a single subfeature to `@import` for the `layer` function. I
added a single feature, though I'm not 100% certain if that's the right
call. There's also a `layer` keyword that creates an anonymous layer.
Conceptually, is this a single function called with or without an
argument? Or is this distinctly two features: a keyword and a function?
I'm not 100% certain and I'd like a review that addresses that question.