Update <link crossorigin> data to match API
Inconsistency pointed out by @gsnedders here:
https://github.com/mdn/browser-compat-data/issues/18368

Most of this is originally from wiki migration:
https://github.com/mdn/browser-compat-data/pull/279

The Chromium implementation was in M34:
https://chromium.googlesource.com/chromium/src/+/21e65a63bf8e3947ff4b4d34cb73866c413106e2
https://www.chromium.org/developers/calendar/

There was no use of "crossorigin" in HTMLLinkElement.cpp before this
point, and https://crbug.com/178787 doesn't say anything to suggest
there was partial support, so assume the old data was wrong.

For Edge and Safari, just trust the data for HTMLLinkElement.