Add missing api.CSS.sv* features
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser.  This particular PR adds the missing `sv*` members of the CSS API, populating the results using data from the [mdn-bcd-collector](https://mdn-bcd-collector.gooborg.com) project (v7.1.4).

Tests Used:
https://mdn-bcd-collector.gooborg.com/tests/api/CSS/svb
https://mdn-bcd-collector.gooborg.com/tests/api/CSS/svh
https://mdn-bcd-collector.gooborg.com/tests/api/CSS/svi
https://mdn-bcd-collector.gooborg.com/tests/api/CSS/svmax
https://mdn-bcd-collector.gooborg.com/tests/api/CSS/svmin
https://mdn-bcd-collector.gooborg.com/tests/api/CSS/svw

_Check out the [collector's guide on how to review this PR](https://github.com/GooborgStudios/mdn-bcd-collector#reviewing-bcd-changes)._