Add missing api.Window.getDigitalGoodsService feature
This PR is a part of a project to add missing interfaces and interface features to BCD that are from an active spec (including WICG specs) and is supported in at least one browser.  This particular PR adds the missing `getDigitalGoodsService` member of the Window API, populating the results using data from ChromeStatus at https://chromestatus.com/feature/5339955595313152.