Add Chromium versions for api.Navigator.share (#18857)
This PR adds real values for Chromium (Chrome, Opera, Samsung Internet, WebView Android) for the `share` member of the `Navigator` API.  I had run some tests in Linux, macOS and Windows on both Chrome and Edge, and determined that non-Windows support for `navigator.share` is still not available.  (See https://crbug.com/770595 for confirmation.)  It appears that this data was updated in #14871,, but it turns out thaat macOS support is only available on Edge, not standard Chrome.  This fixes #16823, fixes #18340 and #18806.