Update some HTML spec URLs to new nav-history-apis.html filename
The multipage HTML spec recently dropped the former history.html file
and added a new nav-history-apis.html file, which includes parts of what
had been in the browsing-the-web.html file and the history.html file.
So, a number of the fragments in BCD spec URLs have moved to the new
nav-history-apis.html file.