Remove HTMLVideoElement»autoPictureInPicture
https://github.com/w3c/picture-in-picture/pull/216 removed
HTMLVideoElement»autoPictureInPicture from the spec, and while the
current BCD data indicates it was implemented in Safari, that’s not
strictly true.

https://github.com/w3c/picture-in-picture/pull/216#issuecomment-1339792930
explains:
> autoPictureInPicture is in WebKit but doesn't actually do anything.
> See https://github.com/search?q=repo%3AWebKit%2FWebKit+autoPictureInPicture&type=code

So it won’t ever actually do anything in Safari, and won’t ever be
implemented in other browsers; so we should remove it from BCD rather
than just it hanging around — because it’s not something that web
developers will ever be able to actually use, anywhere.